Revert "GEODE-9369: Command to copy region entries from a WAN site to another (#6833)" (#7005)
This reverts commit c106e017b2e2ddc07a299bc3d47f6a4baae1451a.

+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/sockets/CacheServerStats.java
@@ -992,8 +992,8 @@ public class CacheServerStats implements MessageStats {
     return this.stats.getDouble(loadPerConnectionId);
   }
 
-  public long getProcessBatchRequests() {
-    return this.stats.getLong(processBatchRequestsId);
+  public int getProcessBatchRequests() {
+    return this.stats.getInt(processBatchRequestsId);
   }
 
   public void close() {
